requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.2
pandas>=2.2.0
openpyxl>=3.1.2
requests>=2.31.0
lxml>=4.9.3
webdriver-manager>=4.0.1
Pillow>=10.1.0
python-dotenv>=1.0.0